fix: created an interface for ProfileRepositoryFirestore and rename function ´getProfile´ to ´fetchProfile´ for architecture consistency - Update the use of the renamed function in the test files
