+++ app/src/main/java/com/epfl/beatlink/model/profile/ProfileRepository.kt
@@ -1,28 +1,9 @@
 package com.epfl.beatlink.model.profile
 
-import android.util.Log
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.toObject
-import kotlinx.coroutines.tasks.await
+interface ProfileRepository {
+    fun getUserId(): String?
 
-class ProfileRepository(private val db: FirebaseFirestore, private val auth: FirebaseAuth) {
-  suspend fun getProfile(userId: String): ProfileData? {
-    return try {
-      val snapshot = db.collection("userProfiles").document(userId).get().await()
-      val data = snapshot.toObject<ProfileData>()
-      Log.d("PROFILE_FETCH", "Fetched profile data: $data") // Added log for debugging
-      data
-    } catch (e: Exception) {
-      e.printStackTrace()
-      Log.e("PROFILE_FETCH_ERROR", "Error fetching profile: ${e.message}") // Log error details
-      null
-    }
-  }
+    suspend fun fetchProfile(userId: String): ProfileData?
 
-  fun getUserId(): String? {
-    val userId = auth.currentUser?.uid
-    Log.d("AUTH", "Current user ID: $userId") // Log user ID for debugging
-    return userId
-  }
+    fun updateProfile(userId: String, profileData: ProfileData): Boolean
 }
